requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.11.0
psycopg==3.2.9
psycopg-binary==3.2.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.33.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typer==0.16.0
types-pyasn1==0.6.0.20250516
types-python-jose==3.5.0.20250531
typing-inspection==0.4.1
typing_extensions==4.14.0
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
